+++ src/BufferedTransport.h
@@ -65,80 +65,80 @@ public:
 	const void * peek(size_t &size);
 	int peek(char *buffer, size_t size);
 	void throwAway(size_t size);
-	int insertInt8(size_t pos, boost::int8_t value);
+	int insertInt8(size_t pos, boost::uint8_t value);
 
-	boost::int32_t writeInt64(boost::int64_t value) {
-		boost::int64_t net;
+	boost::uint32_t writeInt64(boost::uint64_t value) {
+		boost::uint64_t net;
 		if(_setLittleEndian_on) {
-			net = (boost::int64_t)value;
+			net = (boost::uint64_t)value;
 		} else {
 #if BYTE_ORDER == LITTLE_ENDIAN	// TODO this macro is not avaliable on Win32 (need alternatives..)
-			net = ((boost::int64_t)htonl(value) << 32) | ((boost::int64_t)htonl(value >> 32)) ;
+			net = ((boost::uint64_t)htonl(value) << 32) | ((boost::uint64_t)htonl(value >> 32)) ;
 #else
 			net = value;
 #endif
 		}
-		write((boost::int8_t*)&net, 8);
+		write((boost::uint8_t*)&net, 8);
 		return 8;
 	}
 	
-	boost::int32_t writeInt32(boost::int32_t value) {
-		boost::int32_t net;
+	boost::uint32_t writeInt32(boost::uint32_t value) {
+		boost::uint32_t net;
 		if(_setLittleEndian_on) {
-			net = (boost::int32_t)value;
+			net = (boost::uint32_t)value;
 		} else {
-			net = (boost::int32_t)htonl(value);
+			net = (boost::uint32_t)htonl(value);
 		}
-		write((boost::int8_t*)&net, 4);
+		write((boost::uint8_t*)&net, 4);
 		return 4;
 	}
 	
-	boost::int32_t writeInt16(boost::int16_t value) {
-		boost::int16_t net;
+	boost::uint32_t writeInt16(boost::uint16_t value) {
+		boost::uint16_t net;
 		if(_setLittleEndian_on) {
-			net = (boost::int16_t)value;
+			net = (boost::uint16_t)value;
 		} else {
-			net = (boost::int16_t)htons(value);
+			net = (boost::uint16_t)htons(value);
 		}
-		write((boost::int8_t*)&net, 2);
+		write((boost::uint8_t*)&net, 2);
 		return 2;
 	}
 	
-	boost::int32_t writeInt8(boost::int8_t value) {
-		write((boost::int8_t*)&value, 1);
+	boost::uint32_t writeInt8(boost::uint8_t value) {
+		write((boost::uint8_t*)&value, 1);
 		return 1;
 	}
 	
-	boost::int32_t writeBinary(const void *data, size_t size) {
-		boost::int32_t pos = 0;
+	boost::uint32_t writeBinary(const void *data, size_t size) {
+		boost::uint32_t pos = 0;
 		pos += write(data, size);
 		return pos;
 	}
 
-	boost::int32_t writeString8(const std::string &value) {
-		boost::int8_t pos = 0;
+	boost::uint32_t writeString8(const std::string &value) {
+		boost::uint8_t pos = 0;
 		pos += writeInt8(value.size());
 		pos += write(value.c_str(), value.size());
 		return pos;
 	}
 
-	boost::int32_t writeString16(const std::string &value) {
-		boost::int16_t pos = 0;
+	boost::uint32_t writeString16(const std::string &value) {
+		boost::uint16_t pos = 0;
 		pos += writeInt16(value.size());
 		pos += write(value.c_str(), value.size());
 		return pos;
 	}
 
-	boost::int32_t writeString32(const std::string &value) {
-		boost::int32_t pos = 0;
+	boost::uint32_t writeString32(const std::string &value) {
+		boost::uint32_t pos = 0;
 		pos += writeInt32(value.size());
 		pos += write(value.c_str(), value.size());
 		return pos;
 	}
 
-	boost::int32_t writeString32List(const stringlist_t &list) {
-		boost::int32_t pos = 0;
-		pos += writeInt32((boost::int32_t)list.size());
+	boost::uint32_t writeString32List(const stringlist_t &list) {
+		boost::uint32_t pos = 0;
+		pos += writeInt32((boost::uint32_t)list.size());
 		for(size_t i = 0; i < list.size(); i++) {
 			pos += writeString32(list[i]);
 		}
@@ -147,92 +147,92 @@ public:
 
 	// Reading helper method
 public:
-	boost::int32_t readInt32(boost::int32_t &value) {
+	boost::uint32_t readInt32(boost::uint32_t &value) {
 		union bytes {
-			boost::int8_t b[4];
-			boost::int32_t all;
+			boost::uint8_t b[4];
+			boost::uint32_t all;
 		} theBytes;
 
 		if(read((void *)theBytes.b, 4) != 4)
 			throw ProtocolException("readInt32 failed..");
 		else {
 			if(_setLittleEndian_on)
-				value = (boost::int32_t)theBytes.all;
+				value = (boost::uint32_t)theBytes.all;
 			else
-				value = (boost::int32_t)ntohl(theBytes.all);
+				value = (boost::uint32_t)ntohl(theBytes.all);
 		}
 		return 4;
 	}
 
-	boost::int64_t readInt64() {
+	boost::uint64_t readInt64() {
 		union bytes {
-			boost::int8_t b[8];
-			boost::int64_t all;
+			boost::uint8_t b[8];
+			boost::uint64_t all;
 		} theBytes;
 
 		if(read((void *)theBytes.b, 8) != 8)
 			throw ProtocolException("readInt64 failed..");
 		if(_setLittleEndian_on) {
-			return (boost::int64_t)theBytes.all;
+			return (boost::uint64_t)theBytes.all;
 		} else {
 #if BYTE_ORDER == LITTLE_ENDIAN	// TODO this macro is not avaliable on Win32 (need alternatives..)
-			return ((boost::int64_t)(ntohl(theBytes.all >> 32))) | ((boost::int64_t)ntohl(theBytes.all) << 32);
+			return ((boost::uint64_t)(ntohl(theBytes.all >> 32))) | ((boost::uint64_t)ntohl(theBytes.all) << 32);
 #else 
-			return (boost::int64_t)theBytes.all;
+			return (boost::uint64_t)theBytes.all;
 #endif
 		}
 	}
 
 
-	boost::int32_t readInt32() {
+	boost::uint32_t readInt32() {
 		union bytes {
-			boost::int8_t b[4];
-			boost::int32_t all;
+			boost::uint8_t b[4];
+			boost::uint32_t all;
 		} theBytes;
 
 		if(read((void *)theBytes.b, 4) != 4)
 			throw ProtocolException("readInt32 failed..");
 		if(_setLittleEndian_on) {
-			return (boost::int32_t)theBytes.all;
+			return (boost::uint32_t)theBytes.all;
 		} else {
-			return (boost::int32_t)ntohl(theBytes.all);
+			return (boost::uint32_t)ntohl(theBytes.all);
 		}
 	}
 
-	boost::int16_t readInt16(boost::int16_t &value) {
+	boost::uint16_t readInt16(boost::uint16_t &value) {
 		union bytes {
-			boost::int8_t b[2];
-			boost::int16_t all;
+			boost::uint8_t b[2];
+			boost::uint16_t all;
 		} theBytes;
 
 		if(read((void *)theBytes.b, 2) != 2)
 			throw ProtocolException("readInt16 failed..");
 		else {
 			if(_setLittleEndian_on)
-				value = (boost::int16_t)theBytes.all;
+				value = (boost::uint16_t)theBytes.all;
 			else
-				value = (boost::int16_t)ntohs(theBytes.all);
+				value = (boost::uint16_t)ntohs(theBytes.all);
 		}
 		return 2;
 	}
 
-	boost::int16_t readInt16() {
+	boost::uint16_t readInt16() {
 		union bytes {
-			boost::int8_t b[2];
-			boost::int16_t all;
+			boost::uint8_t b[2];
+			boost::uint16_t all;
 		} theBytes;
 
 		if(read((void *)theBytes.b, 2) != 2)
 			throw ProtocolException("readInt16 failed..");
 		if(_setLittleEndian_on) {
-			return (boost::int16_t)theBytes.all;
+			return (boost::uint16_t)theBytes.all;
 		} else {
-			return (boost::int16_t)ntohs(theBytes.all);
+			return (boost::uint16_t)ntohs(theBytes.all);
 		}
 	}
 
-	boost::int8_t readInt8(boost::int8_t &value) {
-		boost::int8_t b[1];
+	boost::uint8_t readInt8(boost::uint8_t &value) {
+		boost::uint8_t b[1];
 		int res = read((void *)b, 1);
 		if(res != 1)
 			throw ProtocolException("readInt8 failed..");
@@ -241,20 +241,20 @@ public:
 		return 1;
 	}
 
-	boost::int8_t readInt8() {
-		boost::int8_t b[1];
+	boost::uint8_t readInt8() {
+		boost::uint8_t b[1];
 		int res = read((void *)b, 1);
 		if(res != 1)
 			throw ProtocolException("readInt8 failed..");
 		return b[0];
 	}
 
-	boost::int32_t readString32(std::string &value) {
-		boost::int32_t len = 0;
-		boost::int8_t pos = 0;
+	boost::uint32_t readString32(std::string &value) {
+		boost::uint32_t len = 0;
+		boost::uint8_t pos = 0;
 		pos += readInt32(len);
 		int nread = read(value, len);
-		if(nread != len)
+		if(nread != (int)len)
 			throw ProtocolException("readString32 failed..");
 		return pos + len;
 	}
@@ -268,9 +268,9 @@ public:
 		return str;
 	}
 
-	boost::int32_t readString16(std::string &value) {
-		boost::int16_t len = 0;
-		boost::int16_t pos = 0;
+	boost::uint32_t readString16(std::string &value) {
+		boost::uint16_t len = 0;
+		boost::uint16_t pos = 0;
 		pos += readInt16(len);
 		int nread = read(value, len);
 		if(nread != len)
@@ -287,9 +287,9 @@ public:
 		return str;
 	}
 
-	boost::int32_t readString8(std::string &value) {
-		boost::int8_t len = 0;
-		boost::int32_t pos = 0;
+	boost::uint32_t readString8(std::string &value) {
+		boost::uint8_t len = 0;
+		boost::uint32_t pos = 0;
 		pos += readInt8(len);
 		int nread = read(value, len);
 		if(nread != len)
